+++ assets/src/components/media/manager/VideoUploadWarning.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+export const VideoUploadWarning = () => (
+  <div className="alert alert-info show" role="alert">
+    File uploads have a 20mb limit. We recommend uploading your video to YouTube and then embedding
+    that within your lesson for most uses.
+  </div>
+);