+++ assets/src/components/video_player/ControlBar.tsx
@@ -1,0 +1,18 @@
+interface VideoPlayer {
+  hasStarted: boolean;
+  paused: boolean;
+  volume: number;
+  muted: boolean;
+  isFullscreen: boolean;
+}
+
+export interface ControlButtonProps {
+  order?: number;
+  actions?: {
+    play: () => void;
+    pause: () => void;
+    mute: (muted: boolean) => void;
+    toggleFullscreen: (player: VideoPlayer) => void;
+  };
+  player?: VideoPlayer;
+}