+++ lib/oli/rendering/content.ex
@@ -21,6 +21,7 @@ defmodule Oli.Rendering.Content do
   @callback h6(%Context{}, next, %{}) :: [any()]
   @callback img(%Context{}, next, %{}) :: [any()]
   @callback img_inline(%Context{}, next, %{}) :: [any()]
+  @callback video(%Context{}, next, %{}) :: [any()]
   @callback youtube(%Context{}, next, %{}) :: [any()]
   @callback iframe(%Context{}, next, %{}) :: [any()]
   @callback audio(%Context{}, next, %{}) :: [any()]
@@ -175,6 +176,9 @@ defmodule Oli.Rendering.Content do
 
       "img_inline" ->
         writer.img_inline(context, next, element)
+
+      "video" ->
+        writer.video(context, next, element)
 
       "youtube" ->
         writer.youtube(context, next, element)