+++ assets/src/apps/delivery/store/features/groups/actions/deck.ts
@@ -29,13 +29,17 @@ import {
   setActivities,
   setCurrentActivityId,
 } from '../../activities/slice';
-import { selectHistoryNavigationActivity, setLessonEnd } from '../../adaptivity/slice';
+import {
+  selectHistoryNavigationActivity,
+  setInitPhaseComplete,
+  setLessonEnd,
+} from '../../adaptivity/slice';
 import { loadActivityAttemptState, updateExtrinsicState } from '../../attempt/slice';
 import {
   selectActivityTypes,
+  selectIsInstructor,
   selectNavigationSequence,
   selectPreviewMode,
-  selectIsInstructor,
   selectResourceAttemptGuid,
   selectSectionSlug,
   setScore,
@@ -48,6 +52,7 @@ import { SequenceBank, SequenceEntry, SequenceEntryType } from './sequence';
 export const initializeActivity = createAsyncThunk(
   `${GroupsSlice}/deck/initializeActivity`,
   async (activityId: ResourceId, thunkApi) => {
+    thunkApi.dispatch(setInitPhaseComplete(false));
     const rootState = thunkApi.getState() as RootState;
     const isPreviewMode = selectPreviewMode(rootState);
     const sectionSlug = selectSectionSlug(rootState);