+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Bug Fixes
 
+- Allow insertion of tables, iframes and all other elements in stem, choices and feedback
+
 ### Enhancements
 
 ## 0.21.2 (2022-07-20)