+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Bug Fixes
 
 - Fix an issue where guest user_id is blank in datashop export
+- Allow insertion of tables, iframes and all other elements in stem, choices and feedback
+
+### Enhancements
 
 ## 0.21.2 (2022-07-20)